moved pytests into api-tests folder
